+++ packages/expo-constants/android/src/main/java/expo/modules/constants/ConstantsService.java
@@ -32,6 +32,22 @@ public class ConstantsService implements InternalModule, ConstantsInterface {
   private static final String PREFERENCES_FILE_NAME = "host.exp.exponent.SharedPreferences";
   private static final String UUID_KEY = "uuid";
 
+  public enum ExecutionEnvironment {
+    BARE("bare"),
+    STANDALONE("standalone"),
+    STORE_CLIENT("storeClient");
+
+    private final String mString;
+
+    ExecutionEnvironment(String string) {
+      mString = string;
+    }
+
+    public String getString() {
+      return mString;
+    }
+  }
+
   private static int convertPixelsToDp(float px, Context context) {
     Resources resources = context.getResources();
     DisplayMetrics metrics = resources.getDisplayMetrics();
@@ -64,6 +80,7 @@ public class ConstantsService implements InternalModule, ConstantsInterface {
     Map<String, Object> constants = new HashMap<>();
 
     constants.put("sessionId", mSessionId);
+    constants.put("executionEnvironment", ExecutionEnvironment.BARE.getString());
     constants.put("statusBarHeight", getStatusBarHeight());
     constants.put("deviceYearClass", getDeviceYearClass());
     constants.put("deviceName", getDeviceName());